Antilife Shell changed
Previously the spell was largely useless, now it can be used to send people near you away from you, push people and creatures away easily, and doesn't require concentration. Sure the original version would have been useful in the D&D 5e universe where you only really fight monsters and soldiers who have melee attacks, but now that we are applying it to the real world and have many abilities that negate the touch spell ranges in the Sorcerer and Warlock classes, we have to bump up the powers of an Antilife shell without fundamentally changing what it is meant to do.