Avoided the Druid becoming like a Paladin
The Druid at level 19 gains the ability to become non-standard creatures via wildshape instead of having a pool of necrotic damage to use on attacks when they hit, and nerfed at level 20 to only have the at-will wild shape apply to CR 2 or lower standard creatures, while having effectively double the wild shapes for CR 4 and 3 standard creatures, and an additional one every 25% of changes for CR 5 and 6 standard creatures, and using the wild shapes as normal for CR 8 and 7 creatures and non-standard creatures (of which the limitations are very clearly stated in terms of only using already existent abilities and having a strict limitation on what the stat blocks can look like (including resistances, immunities AC and hp)). 
Gave an option that makes the game substantially easier to implement in a real-world setting by having each action and multiattack have its own initiative which is resorted at the beginning of the next round of combat, and not having one-action spells enable you to still use your multiattack.