Revised Sorcerer and Death Domain
I realized that the Death Domain Clerics ought not to have an at-will healing spell due to that not fitting, and instead given them the 11th level ability to use Create Undead to make even better and more undead, and allow them to reassert control over all undead with Animate Dead, and have a better consideration for the fact that with the revised timeline for how long a long rest is now (24 hours) and the fact that clerics only get their spell slots back on a full long rest that the duration which their undead are controlled for (previously 24 hours) would need to be increased commensurately. Thus the 11th level Death Domain cleric ability is to have control over undead (via Animate Dead or Create Undead) be a number of days equal to the level of spell slot used, which directly backs up the 5th level ability to further add to the numbers of undead that could be made, further backs up the kind-of Horocrux ability that can be used at 7th level, and further backs up being able to make far more powerful undead such as Ghosts and Wraiths.

Cleric was also revised to give one 10th level spell slot at level 20, to help balance with the Druid's 20th level abilities, and the fact that Warlocks now have feats. Thus, I revised Sorcerer up with a greater numerical edge over all over classes relative to 1st level spells by increasing their Sorcery point maximum more directly with their charisma score (quickly increasing by level) and by increasing the cost to them of high-level slots (sorcery points=ceil(level of spell slot^1.63)) to maintain the (in general) same ability to cast 9th level spells when at 20th level as they would generally be able to under the previous formulation of sorcery points vs charisma and level.