Added Maddening Darkness to Illusion
I also moved around the Mordenkainen's Arcane Sword to Level 5 with the statement that it counts as a level 7 spell for the purpose of getting past Globe of Invulnerability and a Rakshasa (and other creature)'s "limited magic immunity." This is because the spell is severely underpowered as a concentration-single-target spell of high level that really needs something that differentiates it to make it useful to know. This fits with the idea of always creating edge cases to get around particular abilities.
I also added Steel Wind Strike into this since it is a great spell.
I put Whirlwind into the School of Conjuration to fit with Storm of Vengeance also being a Conjuration spell.